add, update, delete user with registration (refactoring)

+++ src/main/java/com/kuzmenchuk/publications/repository/UserRepository.java
@@ -3,10 +3,11 @@ package com.kuzmenchuk.publications.repository;
 import com.kuzmenchuk.publications.repository.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
     User findByUsername(String username);
-    Optional<User> findById(Integer id);
+    Optional<User> findById(Long id);
 }
